feat(config): implement V1 of runtime configuration management
