+++ Cargo.toml
@@ -9,3 +9,4 @@ edition = "2021"
 config = "0.13.1"
 directories = "4.0.1"
 serde = { version = "1.0.137", features = ["derive"] }
+shellexpand = "2.1.0"
