--- config/example.toml
@@ -1,2 +1,0 @@
-[aliasmanager]
-shell_file_path = "$HOME/.zshrc"
